internal/dag: split builder and status tests
Updates #1426
