+++ internal/dag/builder_test.go
@@ -3343,632 +3343,6 @@ func TestMatchesPathPrefix(t *testing.T) {
 	}
 }
 
-func TestDAGIngressRouteStatus(t *testing.T) {
-	// ir1 is a valid ingressroute
-	ir1 := &ingressroutev1.IngressRoute{
-		ObjectMeta: metav1.ObjectMeta{
-			Namespace: "roots",
-			Name:      "example",
-		},
-		Spec: ingressroutev1.IngressRouteSpec{
-			VirtualHost: &ingressroutev1.VirtualHost{
-				Fqdn: "example.com",
-			},
-			Routes: []ingressroutev1.Route{{
-				Match: "/foo",
-				Services: []ingressroutev1.Service{{
-					Name: "home",
-					Port: 8080,
-				}},
-			}, {
-				Match: "/prefix",
-				Delegate: &ingressroutev1.Delegate{
-					Name: "delegated",
-				}},
-			},
-		},
-	}
-
-	// ir2 is invalid because it contains a service with negative port
-	ir2 := &ingressroutev1.IngressRoute{
-		ObjectMeta: metav1.ObjectMeta{
-			Namespace: "roots",
-			Name:      "example",
-		},
-		Spec: ingressroutev1.IngressRouteSpec{
-			VirtualHost: &ingressroutev1.VirtualHost{
-				Fqdn: "example.com",
-			},
-			Routes: []ingressroutev1.Route{{
-				Match: "/foo",
-				Services: []ingressroutev1.Service{{
-					Name: "home",
-					Port: -80,
-				}},
-			}},
-		},
-	}
-
-	// ir3 is invalid because it lives outside the roots namespace
-	ir3 := &ingressroutev1.IngressRoute{
-		ObjectMeta: metav1.ObjectMeta{
-			Namespace: "finance",
-			Name:      "example",
-		},
-		Spec: ingressroutev1.IngressRouteSpec{
-			VirtualHost: &ingressroutev1.VirtualHost{
-				Fqdn: "example.com",
-			},
-			Routes: []ingressroutev1.Route{{
-				Match: "/foobar",
-				Services: []ingressroutev1.Service{{
-					Name: "home",
-					Port: 8080,
-				}},
-			}},
-		},
-	}
-
-	// ir4 is invalid because its match prefix does not match its parent's (ir1)
-	ir4 := &ingressroutev1.IngressRoute{
-		ObjectMeta: metav1.ObjectMeta{
-			Namespace: "roots",
-			Name:      "delegated",
-		},
-		Spec: ingressroutev1.IngressRouteSpec{
-			Routes: []ingressroutev1.Route{{
-				Match: "/doesnotmatch",
-				Services: []ingressroutev1.Service{{
-					Name: "home",
-					Port: 8080,
-				}},
-			}},
-		},
-	}
-
-	// ir5 is invalid because its service weight is less than zero
-	ir5 := &ingressroutev1.IngressRoute{
-		ObjectMeta: metav1.ObjectMeta{
-			Namespace: "roots",
-			Name:      "delegated",
-		},
-		Spec: ingressroutev1.IngressRouteSpec{
-			VirtualHost: &ingressroutev1.VirtualHost{
-				Fqdn: "example.com",
-			},
-			Routes: []ingressroutev1.Route{{
-				Match: "/foo",
-				Services: []ingressroutev1.Service{{
-					Name:   "home",
-					Port:   8080,
-					Weight: -10,
-				}},
-			}},
-		},
-	}
-
-	// ir6 is invalid because it delegates to itself, producing a cycle
-	ir6 := &ingressroutev1.IngressRoute{
-		ObjectMeta: metav1.ObjectMeta{
-			Namespace: "roots",
-			Name:      "self",
-		},
-		Spec: ingressroutev1.IngressRouteSpec{
-			VirtualHost: &ingressroutev1.VirtualHost{
-				Fqdn: "example.com",
-			},
-			Routes: []ingressroutev1.Route{{
-				Match: "/foo",
-				Delegate: &ingressroutev1.Delegate{
-					Name: "self",
-				},
-			}},
-		},
-	}
-
-	// ir7 delegates to ir8, which is invalid because it delegates back to ir7
-	ir7 := &ingressroutev1.IngressRoute{
-		ObjectMeta: metav1.ObjectMeta{
-			Namespace: "roots",
-			Name:      "parent",
-		},
-		Spec: ingressroutev1.IngressRouteSpec{
-			VirtualHost: &ingressroutev1.VirtualHost{
-				Fqdn: "example.com",
-			},
-			Routes: []ingressroutev1.Route{{
-				Match: "/foo",
-				Delegate: &ingressroutev1.Delegate{
-					Name: "child",
-				},
-			}},
-		},
-	}
-
-	ir8 := &ingressroutev1.IngressRoute{
-		ObjectMeta: metav1.ObjectMeta{
-			Namespace: "roots",
-			Name:      "child",
-		},
-		Spec: ingressroutev1.IngressRouteSpec{
-			Routes: []ingressroutev1.Route{{
-				Match: "/foo",
-				Delegate: &ingressroutev1.Delegate{
-					Name: "parent",
-				},
-			}},
-		},
-	}
-
-	// ir9 is invalid because it has a route that both delegates and has a list of services
-	ir9 := &ingressroutev1.IngressRoute{
-		ObjectMeta: metav1.ObjectMeta{
-			Namespace: "roots",
-			Name:      "parent",
-		},
-		Spec: ingressroutev1.IngressRouteSpec{
-			VirtualHost: &ingressroutev1.VirtualHost{
-				Fqdn: "example.com",
-			},
-			Routes: []ingressroutev1.Route{{
-				Match: "/foo",
-				Delegate: &ingressroutev1.Delegate{
-					Name: "child",
-				},
-				Services: []ingressroutev1.Service{{
-					Name: "kuard",
-					Port: 8080,
-				}},
-			}},
-		},
-	}
-
-	// ir10 delegates to ir11 and ir 12.
-	ir10 := &ingressroutev1.IngressRoute{
-		ObjectMeta: metav1.ObjectMeta{
-			Namespace: "roots",
-			Name:      "parent",
-		},
-		Spec: ingressroutev1.IngressRouteSpec{
-			VirtualHost: &ingressroutev1.VirtualHost{
-				Fqdn: "example.com",
-			},
-			Routes: []ingressroutev1.Route{{
-				Match: "/foo",
-				Delegate: &ingressroutev1.Delegate{
-					Name: "validChild",
-				},
-			}, {
-				Match: "/bar",
-				Delegate: &ingressroutev1.Delegate{
-					Name: "invalidChild",
-				},
-			}},
-		},
-	}
-
-	ir11 := &ingressroutev1.IngressRoute{
-		ObjectMeta: metav1.ObjectMeta{
-			Namespace: "roots",
-			Name:      "validChild",
-		},
-		Spec: ingressroutev1.IngressRouteSpec{
-			Routes: []ingressroutev1.Route{{
-				Match: "/foo",
-				Services: []ingressroutev1.Service{{
-					Name: "foo2",
-					Port: 8080,
-				}},
-			}},
-		},
-	}
-
-	// ir12 is invalid because it contains an invalid port
-	ir12 := &ingressroutev1.IngressRoute{
-		ObjectMeta: metav1.ObjectMeta{
-			Namespace: "roots",
-			Name:      "invalidChild",
-		},
-		Spec: ingressroutev1.IngressRouteSpec{
-			Routes: []ingressroutev1.Route{{
-				Match: "/bar",
-				Services: []ingressroutev1.Service{{
-					Name: "foo3",
-					Port: 12345678,
-				}},
-			}},
-		},
-	}
-
-	// ir13 is invalid because it does not specify and FQDN
-	ir13 := &ingressroutev1.IngressRoute{
-		ObjectMeta: metav1.ObjectMeta{
-			Namespace: "roots",
-			Name:      "parent",
-		},
-		Spec: ingressroutev1.IngressRouteSpec{
-			VirtualHost: &ingressroutev1.VirtualHost{},
-			Routes: []ingressroutev1.Route{{
-				Match: "/foo",
-				Services: []ingressroutev1.Service{{
-					Name: "foo",
-					Port: 8080,
-				}},
-			}},
-		},
-	}
-
-	// ir14 delegates tp ir15 but it is invalid because it is missing fqdn
-	ir14 := &ingressroutev1.IngressRoute{
-		ObjectMeta: metav1.ObjectMeta{
-			Namespace: "roots",
-			Name:      "invalidParent",
-		},
-		Spec: ingressroutev1.IngressRouteSpec{
-			VirtualHost: &ingressroutev1.VirtualHost{},
-			Routes: []ingressroutev1.Route{{
-				Match: "/foo",
-				Delegate: &ingressroutev1.Delegate{
-					Name: "validChild",
-				},
-			}},
-		},
-	}
-
-	// ir15 is invalid because it contains a wildcarded fqdn
-	ir15 := &ingressroutev1.IngressRoute{
-		ObjectMeta: metav1.ObjectMeta{
-			Namespace: "roots",
-			Name:      "example",
-		},
-		Spec: ingressroutev1.IngressRouteSpec{
-			VirtualHost: &ingressroutev1.VirtualHost{
-				Fqdn: "example.*.com",
-			},
-			Routes: []ingressroutev1.Route{{
-				Match: "/foo",
-				Services: []ingressroutev1.Service{{
-					Name: "home",
-					Port: 8080,
-				}},
-			}},
-		},
-	}
-
-	// ir16 is invalid because it references an invalid service
-	ir16 := &ingressroutev1.IngressRoute{
-		ObjectMeta: metav1.ObjectMeta{
-			Namespace: "roots",
-			Name:      "invalidir",
-		},
-		Spec: ingressroutev1.IngressRouteSpec{
-			VirtualHost: &ingressroutev1.VirtualHost{
-				Fqdn: "example.com",
-			},
-			Routes: []ingressroutev1.Route{{
-				Match: "/foo",
-				Services: []ingressroutev1.Service{{
-					Name: "invalid",
-					Port: 8080,
-				}},
-			}},
-		},
-	}
-
-	s4 := &v1.Service{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "home",
-			Namespace: "roots",
-		},
-		Spec: v1.ServiceSpec{
-			Ports: []v1.ServicePort{{
-				Name:     "http",
-				Protocol: "TCP",
-				Port:     8080,
-			}},
-		},
-	}
-
-	s5 := &v1.Service{
-		ObjectMeta: metav1.ObjectMeta{
-			Namespace: "roots",
-			Name:      "parent",
-		},
-		Spec: v1.ServiceSpec{
-			Ports: []v1.ServicePort{{
-				Name:     "http",
-				Protocol: "TCP",
-				Port:     8080,
-			}},
-		},
-	}
-
-	s6 := &v1.Service{
-		ObjectMeta: metav1.ObjectMeta{
-			Namespace: "roots",
-			Name:      "foo2",
-		},
-		Spec: v1.ServiceSpec{
-			Ports: []v1.ServicePort{{
-				Name:     "http",
-				Protocol: "TCP",
-				Port:     8080,
-			}},
-		},
-	}
-
-	s7 := &v1.Service{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "foo3",
-			Namespace: "roots",
-		},
-		Spec: v1.ServiceSpec{
-			Ports: []v1.ServicePort{{
-				Name:     "http",
-				Protocol: "TCP",
-				Port:     12345678,
-			}},
-		},
-	}
-
-	tests := map[string]struct {
-		objs []interface{}
-		want []Status
-	}{
-		"valid ingressroute": {
-			objs: []interface{}{ir1, s4},
-			want: []Status{{Object: ir1, Status: "valid", Description: "valid IngressRoute", Vhost: "example.com"}},
-		},
-		"invalid port in service": {
-			objs: []interface{}{ir2},
-			want: []Status{{Object: ir2, Status: "invalid", Description: `route "/foo": service "home": port must be in the range 1-65535`, Vhost: "example.com"}},
-		},
-		"root ingressroute outside of roots namespace": {
-			objs: []interface{}{ir3},
-			want: []Status{{Object: ir3, Status: "invalid", Description: "root IngressRoute cannot be defined in this namespace"}},
-		},
-		"delegated route's match prefix does not match parent's prefix": {
-			objs: []interface{}{ir1, ir4, s4},
-			want: []Status{
-				{Object: ir1, Status: "valid", Description: "valid IngressRoute", Vhost: "example.com"},
-				{Object: ir4, Status: "invalid", Description: `the path prefix "/doesnotmatch" does not match the parent's path prefix "/prefix"`, Vhost: "example.com"},
-			},
-		},
-		"invalid weight in service": {
-			objs: []interface{}{ir5},
-			want: []Status{{Object: ir5, Status: "invalid", Description: `route "/foo": service "home": weight must be greater than or equal to zero`, Vhost: "example.com"}},
-		},
-		"root ingressroute does not specify FQDN": {
-			objs: []interface{}{ir13},
-			want: []Status{{Object: ir13, Status: "invalid", Description: "Spec.VirtualHost.Fqdn must be specified"}},
-		},
-		"self-edge produces a cycle": {
-			objs: []interface{}{ir6},
-			want: []Status{{Object: ir6, Status: "invalid", Description: "route creates a delegation cycle: roots/self -> roots/self", Vhost: "example.com"}},
-		},
-		"child delegates to parent, producing a cycle": {
-			objs: []interface{}{ir7, ir8},
-			want: []Status{
-				{Object: ir7, Status: "valid", Description: "valid IngressRoute", Vhost: "example.com"},
-				{Object: ir8, Status: "invalid", Description: "route creates a delegation cycle: roots/parent -> roots/child -> roots/parent", Vhost: "example.com"},
-			},
-		},
-		"route has a list of services and also delegates": {
-			objs: []interface{}{ir9},
-			want: []Status{{Object: ir9, Status: "invalid", Description: `route "/foo": cannot specify services and delegate in the same route`, Vhost: "example.com"}},
-		},
-		"ingressroute is an orphaned route": {
-			objs: []interface{}{ir8},
-			want: []Status{{Object: ir8, Status: "orphaned", Description: "this IngressRoute is not part of a delegation chain from a root IngressRoute"}},
-		},
-		"ingressroute delegates to multiple ingressroutes, one is invalid": {
-			objs: []interface{}{ir10, ir11, ir12, s6, s7},
-			want: []Status{
-				{Object: ir11, Status: "valid", Description: "valid IngressRoute", Vhost: "example.com"},
-				{Object: ir12, Status: "invalid", Description: `route "/bar": service "foo3": port must be in the range 1-65535`, Vhost: "example.com"},
-				{Object: ir10, Status: "valid", Description: "valid IngressRoute", Vhost: "example.com"},
-			},
-		},
-		"invalid parent orphans children": {
-			objs: []interface{}{ir14, ir11},
-			want: []Status{
-				{Object: ir14, Status: "invalid", Description: "Spec.VirtualHost.Fqdn must be specified"},
-				{Object: ir11, Status: "orphaned", Description: "this IngressRoute is not part of a delegation chain from a root IngressRoute"},
-			},
-		},
-		"multi-parent children is not orphaned when one of the parents is invalid": {
-			objs: []interface{}{ir14, ir11, ir10, s5, s6},
-			want: []Status{
-				{Object: ir14, Status: "invalid", Description: "Spec.VirtualHost.Fqdn must be specified"},
-				{Object: ir11, Status: "valid", Description: "valid IngressRoute", Vhost: "example.com"},
-				{Object: ir10, Status: "valid", Description: "valid IngressRoute", Vhost: "example.com"},
-			},
-		},
-		"invalid FQDN contains wildcard": {
-			objs: []interface{}{ir15},
-			want: []Status{{Object: ir15, Status: "invalid", Description: `Spec.VirtualHost.Fqdn "example.*.com" cannot use wildcards`, Vhost: "example.*.com"}},
-		},
-		"missing service shows invalid status": {
-			objs: []interface{}{ir16},
-			want: []Status{{Object: ir16, Status: "invalid", Description: `Service [invalid:8080] is invalid or missing`, Vhost: ""}},
-		},
-	}
-
-	for name, tc := range tests {
-		t.Run(name, func(t *testing.T) {
-			builder := Builder{
-				Source: KubernetesCache{
-					IngressRouteRootNamespaces: []string{"roots"},
-					FieldLogger:                testLogger(t),
-				},
-			}
-			for _, o := range tc.objs {
-				builder.Source.Insert(o)
-			}
-			dag := builder.Build()
-
-			got := dag.Statuses()
-			if len(tc.want) != len(got) {
-				t.Fatalf("expected:\n%v\ngot\n%v", tc.want, got)
-			}
-
-			for _, ex := range tc.want {
-				var found bool
-				for _, g := range got {
-					if cmp.Equal(ex, g) {
-						found = true
-						break
-					}
-				}
-				if !found {
-					t.Fatalf("expected to find:\n%v\nbut did not find it in:\n%v", ex, got)
-				}
-			}
-		})
-	}
-}
-
-func TestDAGIngressRouteUniqueFQDNs(t *testing.T) {
-	ir1 := &ingressroutev1.IngressRoute{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "example-com",
-			Namespace: "default",
-		},
-		Spec: ingressroutev1.IngressRouteSpec{
-			VirtualHost: &ingressroutev1.VirtualHost{
-				Fqdn: "example.com",
-			},
-			Routes: []ingressroutev1.Route{{
-				Match: "/",
-				Services: []ingressroutev1.Service{{
-					Name: "kuard",
-					Port: 8080,
-				}},
-			}},
-		},
-	}
-
-	// ir2 reuses the fqdn used in ir1
-	ir2 := &ingressroutev1.IngressRoute{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "other-example",
-			Namespace: "default",
-		},
-		Spec: ingressroutev1.IngressRouteSpec{
-			VirtualHost: &ingressroutev1.VirtualHost{
-				Fqdn: "example.com",
-			},
-			Routes: []ingressroutev1.Route{{
-				Match: "/",
-				Services: []ingressroutev1.Service{{
-					Name: "kuard",
-					Port: 8080,
-				}},
-			}},
-		},
-	}
-
-	s1 := &v1.Service{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "kuard",
-			Namespace: "default",
-		},
-		Spec: v1.ServiceSpec{
-			Ports: []v1.ServicePort{{
-				Name:       "http",
-				Protocol:   "TCP",
-				Port:       8080,
-				TargetPort: intstr.FromInt(8080),
-			}},
-		},
-	}
-
-	tests := map[string]struct {
-		objs       []interface{}
-		want       []Vertex
-		wantStatus map[Meta]Status
-	}{
-		"insert ingressroute": {
-			objs: []interface{}{
-				s1, ir1,
-			},
-			want: listeners(
-				&Listener{
-					Port: 80,
-					VirtualHosts: virtualhosts(
-						&VirtualHost{
-							Name: "example.com",
-							routes: routes(
-								prefixroute("/", httpService(s1)),
-							),
-						},
-					),
-				},
-			),
-			wantStatus: map[Meta]Status{
-				{name: ir1.Name, namespace: ir1.Namespace}: {
-					Object:      ir1,
-					Status:      StatusValid,
-					Description: "valid IngressRoute",
-					Vhost:       "example.com",
-				},
-			},
-		},
-		"insert conflicting ingressroutes due to fqdn reuse": {
-			objs: []interface{}{
-				ir1, ir2,
-			},
-			want: []Vertex{},
-			wantStatus: map[Meta]Status{
-				{name: ir1.Name, namespace: ir1.Namespace}: {
-					Object:      ir1,
-					Status:      StatusInvalid,
-					Description: `fqdn "example.com" is used in multiple IngressRoutes: default/example-com, default/other-example`,
-					Vhost:       "example.com",
-				},
-				{name: ir2.Name, namespace: ir2.Namespace}: {
-					Object:      ir2,
-					Status:      StatusInvalid,
-					Description: `fqdn "example.com" is used in multiple IngressRoutes: default/example-com, default/other-example`,
-					Vhost:       "example.com",
-				},
-			},
-		},
-	}
-
-	for name, tc := range tests {
-		t.Run(name, func(t *testing.T) {
-			var builder Builder
-			for _, o := range tc.objs {
-				builder.Source.Insert(o)
-			}
-			dag := builder.Build()
-
-			got := make(map[int]*Listener)
-			dag.Visit(listenerMap(got).Visit)
-
-			want := make(map[int]*Listener)
-			for _, v := range tc.want {
-				if l, ok := v.(*Listener); ok {
-					want[l.Port] = l
-				}
-			}
-
-			opts := []cmp.Option{
-				cmp.AllowUnexported(Listener{}, VirtualHost{}),
-			}
-			if diff := cmp.Diff(want, got, opts...); diff != "" {
-				t.Fatal(diff)
-			}
-
-			gotStatus := dag.statuses
-			if diff := cmp.Diff(tc.wantStatus, gotStatus); diff != "" {
-				t.Fatal(diff)
-			}
-		})
-	}
-}
-
 func TestHttpPaths(t *testing.T) {
 	tests := map[string]struct {
 		rule v1beta1.IngressRule
